Tried to get "Unit" Tests Working
"Unit" tests, because they're really DB integration tests... but, I mean, how
else are we going to test that our stuff works?

Running into some issues getting unit tests working for our DB methods, due to
their asynchronous nature, what with the `Observable`s and such. There appears
to be a way to get it to work, according to these articles:

<https://labs.ribot.co.uk/unit-testing-rxjava-6e9540d4a329>

<http://fedepaol.github.io/blog/2015/09/13/testing-rxjava-observables-subscriptions/>

Included in this commit are some important changes to the `query()` methods in
the currently implemented model classes. Specifically, the returned
`Observable`s are no longer automatically observe the main thread when they are
returned this was necessary to make testing easy, and it will affect the way
that UI integration works.