OneSignal Notification Initial Configuration
The app can now receive push notifications from OneSignal (yes, even in the
emulator!). `NotificationService` is still going to be written on this branch
(see the library API specification for a tentative interface) which will allow
the app itself to send notifications. OneSignal's Android SDK documentation can
be found here:

<https://documentation.onesignal.com/docs/android-native-sdk>