Implemented TableModel.getStatus()
Dear Vivienne and Friends:

tableStatusId is no longer a thing. Instead, just call `.getStatus()` on
whichever TableModel you want the status of -- it'll return an observable. That
does mean if you want the status of all of the tables at once, you're going to
have to call it in a loop. Sorry about that. In a real-world application (maybe
one that wasn't written in the Devil's Tongue, aka Java) this would be a lot
nicer.

When you place an order or make a table reservation, the `.getStatus()` of the
appropriate table will automatically return the correct value when called --
there's nothing extra you have to do to get the table transitioned into the
right state.

Please see the JavaDoc I have written for the `TableModel.getStatus()` method
for extra details on exactly what it returns.

Other commit notes:

The real TableModel status method was never written. Now, at many pains and
frustrations, it is written, and has been tested.

I've also added in some System.out logging to some of the model `query()`
methods that helped me more quickly diagnose the errors I made while designing
this method. These messages were not removed for two chief reasons:

1. They might help someone else, and
2. Even if they don't, I'm too lazy to get rid of them

IMPORTANT: This commit also fixes an issue preventing the `.waiters()`,
`.cooks()`, and `.admins()` methods from working in `RestaurantModel`.